backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
pytest
httpx
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
watchdog
python-frontmatter
